FEAT Added Character Class and fixed typos

+++ src/main/scala/unit/Unit.scala
@@ -4,5 +4,5 @@ trait Unit {
   val name: String
   var life: Int
   var defense: Int
-  val weight: Float
+  val weight: Double
 }
